Replace cli/submit.py with new_submit.py; dispatch its pieces to their modules
`new_submit.py` was the scratch pad while iterating on the top-down rework; now that it
works end-to-end, retire it and move its pieces to where they belong:

- cluv/cli/submit.py: the old (partially broken, mid-refactor) implementation is replaced by
  new_submit.py's orchestration -- Submission/Job/JobRow-based live table, submit(),
  get_submissions(), get_sbatch_command(), merge_sbatch_args(), submit_job() -- plus
  ensure_clean_git_state/sbatch_args_from_dict/build_submit_command/create_submit_commit
  (moved in verbatim from the old submit.py, since new_submit.py depended on importing them
  from it) and the display_commands/raise_on_command_error ContextVars the Hydra launcher
  plugin relies on to force command echoing.
- cluv/cli/submit_utils/chunking.py: apply_chunking() (dict-based) replaces the old
  list-based chunking_update_sbatch_args/get_n_chunks/get_time_from_sbatch_args, which are
  now unused. Restored the SBATCH_TIMELIMIT env-var fallback that new_submit.py's version had
  dropped, so chunking still has all three ways to discover a job's time limit (sbatch args,
  env var, job script header).
- cluv/cli/sync.py: get_cluster_to_remote()/sync_common_part()/sync_per_cluster_part() move
  here, next to the sync steps they call directly. `cluv sync` itself now goes through them
  too instead of the old progress-bar-wrapped sync_task_function/_build_sync_tasks (removed,
  now dead) -- same behavior (dedup for clusters sharing a filesystem, --sync-datasets,
  "not connected to any cluster" error), just without the progress bar, and reusing
  get_cluster_to_remote for cluster resolution instead of duplicating it.
- cluv/cli/submit_utils/first.py: deleted. Fully superseded by the JobRow-based
  wait_for_first_running_job/wait_for_jobs_to_cancel/run_scancel in submit.py.
- cluv/cache.py: cluv.cache.Job is replaced by the Submission/Job pair (Job(Submission) with
  job_id/submitted_at), so there's a single Job type instead of two. save_job/load_jobs
  special-case the two fields that don't survive a straight asdict()/Job(**...) round-trip:
  job_script (Path <-> str) and remote (dropped when persisting -- a stale SSH-multiplexing
  handle isn't meaningful once written to disk; `cluster` already says where the job ran).
  submitted_at stays a real datetime.datetime on the dataclass; only the on-disk JSON uses
  ISO 8601 strings, converted at the save_job/load_jobs boundary.
- cluv/cli/status.py: updated for `job.submitted_at` now being a datetime directly (used to
  be a string requiring `datetime.fromisoformat()` with a fallback for parse failures --
  the fallback was dead code once the type is enforced by the dataclass itself).
- cluv/__main__.py: `cluv submit` now imports from `cli.submit` again instead of
  `cli.new_submit`.

Tests: tests/test_chunking.py and the internals-heavy parts of tests/test_submit.py were
rewritten for the new signatures (dict sbatch_args, no ResolvedSbatchArgs, no separate
submit_first -- cluster="first" is just submit()'s existing race path); the CLI-parsing and
ensure_clean_git_state tests needed no changes since those functions/paths are unchanged.
tests/test_sync_cache.py and tests/test_sync_shared_filesystem.py updated for
sync_task_function -> sync_per_cluster_part (one test that only unit-tested the now-deleted
_build_sync_tasks was dropped; its coverage of the shared-filesystem dedup logic itself
lives on in test_remotes_to_actually_sync_keeps_only_the_first_per_group).

Verified: `ruff check`/`ruff format --check` clean. `pytest -m "not integration"`: 184
passed, only the 3 pre-existing test_hydra_launcher_is_discoverable failures remain (fails
on this machine's temp dir not being under $HOME, independent of this branch -- also called
out as a known environment issue in PR #166's own description). Down from the prior 10
failed/6 errors baseline. `pytest -m integration --collect-only`: collects cleanly (no
import errors), including the submit() integration test that asserts
`isinstance(job, Job)` from cluv.cache -- now true directly, since submit() returns that
type. Hydra plugin (`hydra_plugins.cluv.cluv_launcher`) imports cleanly; its
`job.cluster`/`job.job_id`/`job.n_chunks` accesses after `await submit(...)` are unaffected.
Re-ran the same end-to-end smoke test as the previous commits (mocked sbatch/sacct against
examples/pytorch-example's real config): live table walks
SYNCING -> SUBMITTING -> PENDING -> RUNNING, the job round-trips through save_job/load_jobs
correctly (job_script/submitted_at survive the JSON boundary), and cli/status.py's table
builder consumes the reloaded record without error.
