+++ cluv/cli/submit.py
@@ -1,6 +1,5 @@
-from __future__ import annotations
-
 import asyncio
+import dataclasses
 import datetime
 import logging
 import os
@@ -8,42 +7,422 @@ import shlex
 import subprocess
 import sys
 from contextvars import ContextVar
-from dataclasses import dataclass
 from pathlib import Path, PurePosixPath
-from typing import NamedTuple
 
-import rich.syntax
 import rich.table
 import rich.text
 from rich.live import Live
 
-from cluv.cache import Job, save_job
-from cluv.cli.submit_utils.chunking import chunking_update_sbatch_args, get_n_chunks
-from cluv.cli.submit_utils.first import (
-    cancel_job,
-    wait_for_jobs_to_cancel,
-    wait_for_running_job,
-)
-from cluv.cli.sync import get_active_remotes, pull_datasets_if_needed, run_git_push_if_needed, sync
+from cluv.cache import Job, Submission, save_job
+from cluv.cli.submit_utils.chunking import apply_chunking
+from cluv.cli.sync import get_cluster_to_remote, sync_common_part, sync_per_cluster_part
 from cluv.config import SbatchArgs, find_pyproject, get_cluv_config
 from cluv.remote import Remote, run
-from cluv.utils import console, current_cluster
+from cluv.slurm import FAILED_JOB_STATES, run_saccts
+from cluv.utils import console
 
 logger = logging.getLogger(__name__)
 
 __all__ = ["submit"]
+
 display_commands = ContextVar("display_commands", default=True)
 raise_on_command_error = ContextVar("raise_on_command_error", default=False)
 
+JobState = str
 
-@dataclass(frozen=True)
-class ResolvedSbatchArgs:
-    """The resolved sbatch arguments after merging the config and the command-line args, and the
-    external information we can infer from them, such as the number of chunks for chunked jobs
-    when calling get_sbatch_command()."""
 
-    sbatch_args: list[str]
-    n_chunks: int | None = None
+class JobSubmissionFailed(Exception):
+    """Raised when a job submission fails."""
+
+
+@dataclasses.dataclass
+class JobRow:
+    """Live, mutable view of one `Submission`'s progress.
+
+    Tracks a submission from before it's even synced (``state="SYNCING"``, no job id yet),
+    through submission (``job_id`` known, state polled from `sacct`), to running and,
+    possibly, cancellation.
+
+    A single flat list of these -- covering every submission, on every cluster, for one
+    `submit()` call -- is all a live display needs to render the whole picture, from a plain
+    `rich.Live` table (see `render_job_table` below) up to, eventually, a table shared across
+    several concurrent `submit`/`submit_first` calls (`rich.Live` only supports one live region
+    per console, so that would fuse several such lists together instead of replacing this one).
+    """
+
+    submission: Submission
+    state: JobState = "SYNCING"
+    job_id: int | None = None
+    submitted_at: datetime.datetime | None = None
+    error: str | None = None
+
+    @property
+    def cluster(self) -> str:
+        return self.submission.cluster
+
+    def as_job(self) -> Job | None:
+        """A `Job` snapshot of this row, once it has actually been submitted."""
+        if self.job_id is None or self.submitted_at is None:
+            return None
+        return Job(
+            cluster=self.submission.cluster,
+            remote=self.submission.remote,
+            job_script=self.submission.job_script,
+            sbatch_args=self.submission.sbatch_args,
+            program_args=self.submission.program_args,
+            sbatch_command=self.submission.sbatch_command,
+            n_chunks=self.submission.n_chunks,
+            git_commit=self.submission.git_commit,
+            job_id=self.job_id,
+            submitted_at=self.submitted_at,
+        )
+
+
+def _state_style(state: JobState) -> str:
+    if state.startswith(("RUNNING", "COMPLETED", "CANCELLED")):
+        return "green"
+    if state.startswith(("SYNCING", "SUBMITTING", "PENDING", "UNKNOWN")):
+        return "yellow"
+    return "red"
+
+
+def render_job_table(rows: list[JobRow], *, cancelling: bool = False) -> rich.table.Table:
+    """Render the current state of every submission as a single table.
+
+    A plain `rich.Live` for now; the natural place to plug in a registry that fuses several
+    concurrent `submit`/`submit_first` calls' rows into one shared live region later on.
+    """
+    title = "Waiting for jobs to cancel..." if cancelling else "Submitting jobs..."
+    table = rich.table.Table("Cluster", "Job ID", "Status", title=title)
+    for row in rows:
+        table.add_row(
+            row.cluster,
+            str(row.job_id) if row.job_id is not None else "-",
+            rich.text.Text(row.state, style=_state_style(row.state)),
+        )
+    return table
+
+
+def _group_by_cluster(rows: list[JobRow]) -> dict[str, list[JobRow]]:
+    grouped: dict[str, list[JobRow]] = {}
+    for row in rows:
+        grouped.setdefault(row.cluster, []).append(row)
+    return grouped
+
+
+async def submit(
+    cluster: str,
+    job_script: Path | None,
+    sbatch_args: list[str],
+    program_args: list[str],
+    autocommit: bool = False,
+    chunking: int | None = None,
+    _skip_sync: bool = False,
+) -> Job | None:
+    """Submit a job to the given cluster (or all clusters if `cluster=="first"`),
+    and return the Job object if successful.
+
+    Returns None if the submission failed.
+    """
+    submit_command = build_submit_command(
+        cluster=cluster, job_script=job_script, sbatch_args=sbatch_args, program_args=program_args
+    )
+    git_commit = ensure_clean_git_state(autocommit=autocommit, submit_command=submit_command)
+    cluster_to_remote = await get_cluster_to_remote(cluster)
+
+    rows = [
+        JobRow(submission=submission)
+        for cluster_name, remote in cluster_to_remote.items()
+        for submission in get_submissions(
+            cluster_name,
+            remote,
+            job_script=job_script,
+            sbatch_args=sbatch_args,
+            program_args=program_args,
+            chunking=chunking,
+            git_commit=git_commit,
+        )
+    ]
+
+    if not _skip_sync:
+        await sync_common_part(list(cluster_to_remote.values()))
+
+    found_running_job = asyncio.Event()
+    tasks = [
+        asyncio.create_task(
+            submit_to_cluster(
+                cluster_name,
+                remote,
+                rows=[row for row in rows if row.cluster == cluster_name],
+                found_running_job=found_running_job,
+                _skip_sync=_skip_sync,
+            )
+        )
+        for cluster_name, remote in cluster_to_remote.items()
+    ]
+
+    cancelling = False
+
+    def _render() -> rich.table.Table:
+        return render_job_table(rows, cancelling=cancelling)
+
+    with Live(get_renderable=_render, console=console, refresh_per_second=1):
+        first_running_row = await wait_for_first_running_job(
+            rows, cluster_to_remote, tasks, found_running_job
+        )
+        if first_running_row is None:
+            console.log("All job submissions have failed! Exiting.")
+            return None
+
+        cancelling = True
+        other_rows = [
+            row for row in rows if row is not first_running_row and row.job_id is not None
+        ]
+        await wait_for_jobs_to_cancel(other_rows, cluster_to_remote)
+
+    console.print(
+        f"[green]Job {first_running_row.job_id} on cluster {first_running_row.cluster} is "
+        f"running.[/green]"
+    )
+    job = first_running_row.as_job()
+    assert job is not None
+    save_job(job)
+    return job
+
+
+async def wait_for_first_running_job(
+    rows: list[JobRow],
+    cluster_to_remote: dict[str, Remote | None],
+    tasks: list[asyncio.Task],
+    found_running_job: asyncio.Event,
+    max_wait_time_seconds: int = 60,
+) -> JobRow | None:
+    """Poll `sacct` until one submitted job starts running, or every submission has failed.
+
+    Mutates `rows` in place with the latest known job id / state, so a live display can render
+    them at any point during this wait. Sets `found_running_job` the moment a job starts, so
+    that clusters which haven't submitted their own jobs yet can skip doing so.
+
+    Returns the row for the job that started, or None if every submission ended up failing.
+    """
+    delay = 1
+    while True:
+        submitted = [row for row in rows if row.job_id is not None]
+        by_cluster = _group_by_cluster(submitted)
+        if by_cluster:
+            states_per_cluster = await asyncio.gather(
+                *(
+                    run_saccts(cluster_to_remote[cluster], [row.job_id for row in cluster_rows])
+                    for cluster, cluster_rows in by_cluster.items()
+                )
+            )
+            for cluster_rows, states in zip(by_cluster.values(), states_per_cluster):
+                for row, state in zip(cluster_rows, states):
+                    row.state = state
+                    if state.startswith(("RUNNING", "COMPLETED")):
+                        found_running_job.set()
+                        return row
+
+        all_failed = bool(submitted) and all(
+            row.state.startswith(tuple(FAILED_JOB_STATES)) for row in submitted
+        )
+        if all(task.done() for task in tasks) and (not submitted or all_failed):
+            return None
+
+        await asyncio.sleep(delay)
+        delay = min(delay * 2, max_wait_time_seconds)
+
+
+async def wait_for_jobs_to_cancel(
+    rows: list[JobRow],
+    cluster_to_remote: dict[str, Remote | None],
+    max_wait_time_seconds: int = 60,
+) -> None:
+    """Cancel every (already-submitted) job in `rows`, and wait until they're all done."""
+    to_cancel = [row for row in rows if not row.state.startswith(("CANCELLED", "COMPLETED"))]
+    if not to_cancel:
+        return
+
+    await run_scancel(to_cancel)
+
+    delay = 1
+    while to_cancel:
+        by_cluster = _group_by_cluster(to_cancel)
+        states_per_cluster = await asyncio.gather(
+            *(
+                run_saccts(cluster_to_remote[cluster], [row.job_id for row in cluster_rows])
+                for cluster, cluster_rows in by_cluster.items()
+            )
+        )
+        for cluster_rows, states in zip(by_cluster.values(), states_per_cluster):
+            for row, state in zip(cluster_rows, states):
+                if state.startswith("CANCELLED") or state == "FAILED":
+                    # "CANCELLED by <uid>", and a stray "FAILED" job step on some clusters
+                    # (while the rest of the job is "CANCELLED"), both just mean cancelled.
+                    state = "CANCELLED"
+                row.state = state
+
+        to_cancel = [
+            row for row in to_cancel if not row.state.startswith(("CANCELLED", "COMPLETED"))
+        ]
+        if to_cancel:
+            await asyncio.sleep(delay)
+            delay = min(delay * 2, max_wait_time_seconds)
+
+    console.log(f"Cancelled {len(rows)} other job submission(s).")
+
+
+async def run_scancel(rows: list[JobRow]) -> None:
+    """Cancel the (already-submitted) jobs behind `rows`, grouped by remote."""
+    if not rows:
+        return
+    by_remote: dict[Remote | None, list[JobRow]] = {}
+    for row in rows:
+        by_remote.setdefault(row.submission.remote, []).append(row)
+
+    async def cancel(remote: Remote | None, cluster_rows: list[JobRow]) -> None:
+        job_ids = [row.job_id for row in cluster_rows]
+        scancel_command = f"scancel {' '.join(map(str, job_ids))}"
+        if remote is not None:
+            await remote.get_output(scancel_command, hide=True)
+        else:
+            await run(tuple(shlex.split(scancel_command)), hide=True)
+
+    await asyncio.gather(
+        *(cancel(remote, cluster_rows) for remote, cluster_rows in by_remote.items())
+    )
+
+
+async def submit_to_cluster(
+    cluster: str,
+    remote: Remote | None,
+    rows: list[JobRow],
+    found_running_job: asyncio.Event,
+    _skip_sync: bool = False,
+) -> None:
+    """Sync then submit every submission for one cluster, in parallel."""
+    if not _skip_sync:
+        await sync_per_cluster_part(remote)
+
+    if found_running_job.is_set():
+        # If a job has already started on another cluster, we don't need to submit more jobs.
+        console.log(
+            f"Skipping submission of jobs to cluster {cluster} because a job "
+            f"has already started on another cluster."
+        )
+        for row in rows:
+            row.state = "SKIPPED"
+        return
+
+    for row in rows:
+        row.state = "SUBMITTING"
+
+    results = await asyncio.gather(
+        *(submit_job(row.submission) for row in rows),
+        return_exceptions=True,
+    )
+
+    assert len(results) == len(rows)
+    for row, result in zip(rows, results):
+        if isinstance(result, Job):
+            row.job_id = result.job_id
+            row.submitted_at = result.submitted_at
+            row.state = "PENDING"
+        elif isinstance(result, JobSubmissionFailed):
+            row.error = str(result)
+            row.state = "FAILED"
+            console.log(f"[red]{result}[/red]")
+        else:
+            assert isinstance(result, BaseException)
+            raise result
+
+
+def get_submissions(
+    cluster: str,
+    remote: Remote | None,
+    *,
+    job_script: Path | None,
+    sbatch_args: list[str],
+    program_args: list[str],
+    chunking: int | None,
+    git_commit: str,
+) -> list[Submission]:
+    """Expand the possible job configurations for a cluster. Returns a list of `Submission` objects.
+
+    Does *not* do the actual job submission with `sbatch`.
+    """
+    submissions: list[Submission] = []
+    config = get_cluv_config()
+    cluster_config = config.get_cluster_config(cluster)
+    job_resources_options = cluster_config.sbatch_args
+
+    if job_script is None:
+        if cluster_config.job_script_path is None:
+            raise ValueError(
+                f"No job script specified for cluster {cluster!r}, and no default job script "
+                f"path set in the config."
+            )
+        job_script = Path(os.path.expandvars(str(cluster_config.job_script_path)))
+    if not job_script.exists():
+        raise ValueError(
+            f"The job script ({job_script}) does not exist on this machine. Even though it "
+            f"can be customized per cluster, it needs to exist locally, since cluv needs to "
+            f"read its header to infer sbatch defaults."
+        )
+
+    for job_resources in job_resources_options:
+        job_sbatch_args: SbatchArgs = merge_sbatch_args(
+            from_config=job_resources, from_cli=sbatch_args
+        )
+        n_chunks, job_sbatch_args = apply_chunking(
+            job_sbatch_args, job_script=job_script, chunking=chunking
+        )
+        sbatch_command = get_sbatch_command(
+            cluster,
+            job_script=job_script,
+            sbatch_args=job_sbatch_args,
+            program_args=program_args,
+            git_commit=git_commit,
+        )
+        submissions.append(
+            Submission(
+                cluster=cluster,
+                remote=remote,
+                job_script=job_script,
+                sbatch_args=job_sbatch_args,
+                program_args=program_args,
+                sbatch_command=sbatch_command,
+                n_chunks=n_chunks,
+                git_commit=git_commit,
+            )
+        )
+    return submissions
+
+
+def merge_sbatch_args(from_config: SbatchArgs, from_cli: list[str]) -> SbatchArgs:
+    """Merge the sbatch args from the config and from the CLI, with CLI args taking precedence."""
+    merged = dict(from_config)
+    index = 0
+    while index < len(from_cli):
+        flag = from_cli[index]
+        if flag.startswith("--"):
+            key, _, value = flag[2:].partition("=")
+            merged[key] = value if value else True
+            index += 1
+        elif flag.startswith("-"):
+            key = flag[1:]
+            has_separate_value = index + 1 < len(from_cli) and not from_cli[index + 1].startswith(
+                "-"
+            )
+            if has_separate_value:
+                merged[key] = from_cli[index + 1]
+                index += 2
+            else:
+                merged[key] = True
+                index += 1
+        else:
+            raise ValueError(f"Not a valid sbatch flag: {flag!r}")
+    return merged
 
 
 def sbatch_args_from_dict(d: SbatchArgs) -> list[str]:
@@ -82,461 +461,117 @@ def sbatch_args_from_dict(d: SbatchArgs) -> list[str]:
     return flags
 
 
-class Submission(NamedTuple):
-    """One job to submit: on which cluster, over which connection, and with which sbatch flags."""
-
-    cluster: str
-
-    remote: Remote | None
-    """Remote used to run `sbatch`, or `None` to run it on the current cluster."""
-
-    job_script: Path
-
-    sbatch_args: SbatchArgs
-    """The sbatch flags from the config to use, i.e. one of the cluster's allocations."""
-
-    program_args: list[str]
-
-    sbatch_command: str
-
-    num_chunks: int | None
-
-
-async def submit(
+def get_sbatch_command(
     cluster: str,
-    job_script: Path | None,
-    sbatch_args: list[str],
-    program_args: list[str],
-    autocommit: bool = False,
-    chunking: int | None = None,
-    _skip_sync: bool = False,
-) -> Job | None:
-    """Submit a SLURM job on a remote cluster.
-
-    Enforces a clean git state, syncs the project, sets `GIT_COMMIT` and any
-    environment variables configured in `[tool.cluv.env]` / `[tool.cluv.clusters.<name>.env]`,
-    then calls `sbatch` on the remote.
-
-    `sbatch_args` are forwarded as flags to `sbatch`; `program_args` are passed to
-    the job script.
-
-    When more than one allocation is configured for the target cluster (a list of flag sets in
-    `[tool.cluv.clusters.<name>].sbatch_args`), one job is submitted per allocation and only the
-    first one to start is kept, the others being cancelled.
-
-    Parameters:
-        cluster: SSH hostname of the target cluster. Can be set to "first" to launch the job on all clusters and keep only the first one to starts.
-        job_script: Path to the job script to submit, relative to the project root.
-            When omitted, uses the configured default for the target cluster.
-        sbatch_args: List of additional flags to pass to `sbatch`.
-        program_args: List of arguments to pass to the job script, for example `["python", "main.py"]`.
-        autocommit: If True, automatically create a local commit with tracked changes before submitting.
-        chunking: Duration in hours of each chunk when splitting the job into multiple consecutive
-            short jobs. When None, chunking is disabled.
-        _skip_sync: If True, skip the synchronization step before submitting.
-
-    Returns:
-        The job ID of the submitted job or None if the sbatch command fails.
-
-    Examples:
-
-    ```python
-    submit(
-        "mila",
-        "scripts/job.sh",
-        sbatch_args=["--time=00:00:10"],
-        program_args=["python", "--version"],
-    )
-    ```
-    """
-
-    # Check git is clean locally (untracked files are fine) and capture current commit hash.
-    git_commit = ensure_clean_git_state(
-        autocommit=autocommit,
-        submit_command=build_submit_command(
-            cluster,
-            job_script
-            or (get_job_script_path_from_config(cluster) if cluster != "first" else "")
-            or "<job_script>",
-            sbatch_args,
-            program_args,
-        ),
-    )
-
-    here = current_cluster()
-
-    if cluster == "first":
-        job = await submit_first(
-            job_script, sbatch_args, program_args, git_commit, chunking, _skip_sync=_skip_sync
-        )
-        if job:
-            save_job(job)
-        return job
-
-    if job_script is None:
-        job_script_from_config = get_job_script_path_from_config(cluster)
-        job_script = _check_job_script_exists_locally(job_script_from_config, cluster)
-    else:
-        job_script = _check_job_script_exists_locally(job_script, cluster)
-
-    if cluster != here:
-        # The sbatch command will be run over SSH.
-        if _skip_sync:
-            remote = await Remote.connect(hostname=cluster)
-        else:
-            remote = (await sync(clusters=[cluster]))[0]
-    else:
-        # Submitting to the current cluster. The sbatch command will run locally.
-        remote = None
-
-    job_resources_options = get_cluv_config().get_cluster_config(cluster).sbatch_args
-    if len(job_resources_options) > 1:
-        # We can submit the job using different job configurations, for example
-        # with different accounts, or different GPU models, etc.
-        # Submit one job per allocation and keep the first one that starts.
-        job = await submit_and_keep_first(
-            {
-                remote: [
-                    Submission(
-                        remote=remote,
-                        job_script=job_script,
-                        sbatch_args=sbatch_args_from_dict(job_resources_config) + sbatch_args,
-                        program_args=program_args,
-                    )
-                    for job_resources_config in job_resources_options
-                ]
-            },
-            git_commit=git_commit,
-            chunking=chunking,
-            sync_datasets=True,  # TODO: might be good to have control over this.
-            _skip_sync=False,
-            _skip_sync_common_part=False,
-        )
-        if job:
-            save_job(job)
-        return job
-
-    sbatch_args_from_config = job_resources_options[0]
-    sbatch_args = sbatch_args_from_dict(sbatch_args_from_config) + sbatch_args
-    result, job = await sbatch(
-        remote=remote,
-        job_script=job_script,
-        sbatch_args=sbatch_args,
-        program_args=program_args,
-        git_commit=git_commit,
-        chunking=chunking,
-    )
-
-    if result.returncode != 0 or job is None:
-        console.print(f"[red] Error during sbatch : {result.stderr}[/red]")
-    else:
-        save_job(job)
-
-        console.log(
-            f"Successfully submitted job {job.job_id} on the {cluster} cluster.\n"
-            f"Use `ssh {cluster} sacct -j {job.job_id}` to view its status, and `cluv sync {cluster}` to "
-            f"fetch results once it is complete."
-        )
-
-    return job
-
-
-async def submit_first(
-    job_script: Path | None,
-    sbatch_args: list[str],
+    job_script: Path,
+    sbatch_args: SbatchArgs,
     program_args: list[str],
     git_commit: str,
-    chunking: int | None,
-    _skip_sync: bool = False,
-) -> Job | None:
-    """Submit the job on all clusters (and on every allocation of each cluster), and wait until one
-    of them starts. Once one starts, cancel the others.
+) -> str:
+    """Generate the command to submit the job via `sbatch` on `cluster`, with the appropriate
+    sbatch arguments, environment variables and paths set.
     """
-    remotes = await get_active_remotes()
-    cluster_to_remote: dict[str, Remote | None] = {remote.hostname: remote for remote in remotes}
+    local_project_dir = find_pyproject().parent
+    local_job_script = job_script if job_script.is_absolute() else local_project_dir / job_script
+    job_script_relative_path = local_job_script.relative_to(local_project_dir)
 
-    this_cluster = current_cluster()
-    if this_cluster is not None:
-        # We are also on a Slurm cluster, so consider this as an option as well.
-        cluster_to_remote[this_cluster] = None
-
-    job_scripts = {
-        cluster: _check_job_script_exists_locally(
-            job_script or get_job_script_path_from_config(cluster), cluster
-        )
-        for cluster in cluster_to_remote
-    }
-
-    # One submission per allocation of each cluster.
     config = get_cluv_config()
-    submissions = {
-        remote: [
-            Submission(
-                remote=remote,
-                job_script=job_scripts[cluster],
-                sbatch_args=sbatch_args_from_dict(config_sbatch_args) + sbatch_args,
-                program_args=program_args,
-            )
-            for config_sbatch_args in config.get_cluster_config(cluster).sbatch_args
-        ]
-        for cluster, remote in cluster_to_remote.items()
-    }
-
-    return await submit_and_keep_first(
-        submissions,
-        git_commit=git_commit,
-        chunking=chunking,
-        # todo: might want to add an argument to control this? For example, when submitting
-        # a job from a laptop, but you know the dataset was already synced between clusters?
-        sync_datasets=True,
-        _skip_sync=_skip_sync,
+    cluster_config = config.get_cluster_config(cluster)
+    remote_project_dir = cluster_config.project_dir or (
+        PurePosixPath("$HOME") / local_project_dir.relative_to(Path.home())
     )
+    remote_job_script = PurePosixPath(remote_project_dir) / job_script_relative_path
 
+    env_vars: dict[str, str] = {**config.env, **cluster_config.env}
+    base_name = env_vars.get("SBATCH_JOB_NAME") or Path(job_script).stem
+    env_vars["SBATCH_JOB_NAME"] = f"cluv-{base_name}"
+    env_vars["GIT_COMMIT"] = git_commit
 
-async def sync_and_sbatch(
-    remote: Remote | None,
-    submissions: list[Submission],
-    chunking: int | None,
-    git_commit: str,
-    sync_datasets: bool,
-    _skip_sync: bool = False,
-):
-    # Should all be submissions to the same cluster.
-    if remote is None:
-        here = current_cluster()
-        assert all(submission.cluster == here for submission in submissions)
-    else:
-        assert all(submission.cluster == remote.hostname for submission in submissions)
-
-    if not _skip_sync and remote is not None:
-        await sync([remote.hostname], sync_datasets=sync_datasets, _skip_common_part=True)
-
-    return await asyncio.gather(
-        *[
-            sbatch(
-                submission.remote,
-                job_script=submission.job_script,
-                sbatch_args=submission.sbatch_args,
-                program_args=submission.program_args,
-                git_commit=git_commit,
-                chunking=chunking,
+    sbatch_flags = sbatch_args_from_dict(sbatch_args)
+    if not any("--output" in flag for flag in sbatch_flags):
+        # Chunked (job array) jobs need `%A`/`%a` (array job id / task id) instead of `%j`.
+        if "array" in sbatch_args:
+            env_vars["SBATCH_OUTPUT"] = (
+                f"{cluster_config.results_path}/{cluster}_%A/slurm-%A_%a.out"
             )
-            for submission in submissions
-        ]
-    )
-
-
-async def submit_and_keep_first(
-    submissions: dict[Remote | None, list[Submission]],
-    git_commit: str,
-    chunking: int | None,
-    sync_datasets: bool,
-    _skip_sync: bool = False,
-    _skip_sync_common_part: bool = False,
-) -> Job | None:
-    """Submit all the given jobs, wait until one of them starts, then cancel the others."""
-    cluster_to_remote: dict[str, Remote | None] = {}
-    for remote in submissions.keys():
-        if remote:
-            cluster_to_remote[remote.hostname] = remote
         else:
-            here = current_cluster()
-            assert here is not None
-            cluster_to_remote[here] = None
+            env_vars["SBATCH_OUTPUT"] = f"{cluster_config.results_path}/{cluster}_%j/slurm-%j.out"
 
-    # cluster_task_groups = {cluster: asyncio.TaskGroup() for cluster in cluster_to_remote.keys()}
-
-    # Do the common part of `sync` that should be done for all clusters, and then use `_skip_common_part=True` in each `sync` call below.
-    if not (_skip_sync or _skip_sync_common_part):
-        await run_git_push_if_needed()
-        if sync_datasets:
-            _remotes = [remote for remote in submissions.keys() if remote is not None]
-            await pull_datasets_if_needed(current_cluster(), get_cluv_config(), _remotes)
-
-    cluster_tasks = {
-        remote: asyncio.create_task(
-            sync_and_sbatch(
-                remote,
-                cluster_submissions,
-                chunking=chunking,
-                git_commit=git_commit,
-                sync_datasets=sync_datasets,
+        header_output = next(
+            (
+                line
+                for line in local_job_script.read_text().splitlines()
+                if line.strip().startswith("#SBATCH") and "--output" in line
             ),
-            name=remote.hostname if remote else current_cluster(),
+            None,
         )
-        for remote, cluster_submissions in submissions.items()
-    }
+        if header_output is not None:
+            logger.warning(
+                f"[yellow]The job script {job_script} sets {header_output.strip()!r}, which "
+                f"will be overridden by cluv's SBATCH_OUTPUT so that results can be synced "
+                f"back. Consider using cluv in your Python script to decide where to store "
+                f"results instead.[/yellow]"
+            )
 
-    job_to_state: dict[Job, str] = {}
+    env_vars_prefix = " ".join(f"{k}={shlex.quote(str(v))}" for k, v in env_vars.items())
+    sbatch_args_str = shlex.join(sbatch_flags)
+    program_args_str = shlex.join(program_args)
 
-    def print_cluster_job_submission_results(
-        task: asyncio.Task[list[tuple[subprocess.CompletedProcess[str], Job | None]]],
-    ):
-        cluster = task.get_name()
+    return (
+        f"bash --login -c '{env_vars_prefix} sbatch --parsable --chdir={remote_project_dir} "
+        f"{sbatch_args_str} {remote_job_script} {program_args_str}'"
+    )
 
-        remote = cluster_to_remote[cluster]
-        cluster_submissions = submissions[remote if cluster != current_cluster() else None]
-        # Only show the allocation of each job when some cluster has more than one.
-        show_allocations = len(cluster_submissions) > 1
 
-        table = rich.table.Table(
-            "Cluster",
-            *(["sbatch arguments"] if show_allocations else []),
-            "Result",
-            title="Jobs submitted on the clusters",
+async def submit_job(submission: Submission) -> Job:
+    """Does the actual sbatch call.
+
+    Raise a `JobSubmissionFailed` if the job submission fails for some reason.
+    """
+    display = display_commands.get()
+    hide = not display
+    warn = not raise_on_command_error.get()
+
+    if submission.remote is not None:
+        result = await submission.remote.run(
+            submission.sbatch_command, display=display, warn=warn, hide=hide
         )
-        # Reconstruct the sbatch command that was used, just to display it (not great).
-        try:
-            sbatch_results = task.result()
-        except asyncio.CancelledError as err:
-            console.print(f"One of the job submissions on cluster {cluster} was cancelled.")
-            sbatch_results = [err] * len(cluster_submissions)
-        except Exception as sbatch_exception:
-            # NOTE: Shouldn't really happen, because `sbatch` doesn't raise in case of typical sbatch errors.
-            sbatch_results = [sbatch_exception] * len(cluster_submissions)
-            console.print(f"One of the job submissions on cluster {cluster} failed.")
-
-        for submission, sbatch_result in zip(cluster_submissions, sbatch_results):
-            # todo: Move this `sbatch_command` to a field of `Submission`, since it has to
-            # be computed in `sbatch` anyway, might as well save it.
-            sbatch_command, _sbatch_args = get_sbatch_command(
-                cluster=submission.cluster,
-                job_script=submission.job_script,
-                sbatch_args=submission.sbatch_args,
-                program_args=submission.program_args,
-                git_commit=git_commit,
-                chunking=chunking,
-            )
-            if isinstance(sbatch_result, BaseException):
-                output_text = rich.text.Text(f"Error: {sbatch_result}", style="red")
-            else:
-                result, job = sbatch_result
-                if result.returncode != 0 or job is None:
-                    output_text = rich.text.Text(f"Error: {result.stderr.strip()}", style="red")
-                else:
-                    output_text = rich.text.Text(f"Job ID: {job.job_id}", style="green")
-                    job_to_state[job] = "UNKNOWN"
-            table.add_row(
-                submission.cluster,
-                *([shlex.join(submission.sbatch_args)] if show_allocations else []),
-                rich.console.Group(
-                    rich.syntax.Syntax(sbatch_command, lexer="sh", word_wrap=True),
-                    output_text,
-                ),
-                end_section=True,
-            )
-        console.print(table)
-
-    for cluster, cluster_task in cluster_tasks.items():
-        cluster_task.add_done_callback(print_cluster_job_submission_results)
-
-    # if not job_to_state:
-    #     console.print("No job submitted on clusters. See errors above.")
-    #     return None
-
-    # Wait for a job to start on a cluster.
-    # If the wait is interrupted, cancel all jobs.
-    first_running_job: Job | None = None
-
-    max_wait_time_seconds = 5
-
-    cancelling = False
-
-    def make_table() -> rich.table.Table:
-        table = rich.table.Table(
-            "Cluster",
-            "Job ID",
-            "Status",
-            title="Waiting for a job to start..."
-            if not cancelling
-            else "Waiting for jobs to cancel...",
+    else:
+        result = await run(
+            tuple(shlex.split(submission.sbatch_command)), _display=display, warn=warn, hide=hide
         )
-        for job, state in job_to_state.items():
-            table.add_row(
-                job.cluster,
-                str(job.job_id),
-                rich.text.Text(
-                    state,
-                    style="green"
-                    if state.startswith(("RUNNING", "COMPLETED", "CANCELLED"))
-                    else "yellow"
-                    if state.startswith(("PENDING", "UNKNOWN"))
-                    else "red",
-                ),
-            )
-        return table
 
-    while not any(task.done() for task in cluster_tasks.values()):
-        logger.debug("Waiting for one of the sync+sbatch tasks to complete...")
-        await asyncio.sleep(1)
-
-    try:
-        with Live(get_renderable=make_table, console=console, refresh_per_second=1) as live:
-            wait_result = await wait_for_running_job(
-                job_to_state, cluster_to_remote, max_wait_time_seconds
-            )
-            live.update(make_table(), refresh=True)  # probably not entirely necessary.
-
-            while not wait_result:
-                # If all tasks are done, but no job has started, then all jobs have failed.
-                if all(task.done() for task in cluster_tasks.values()):
-                    console.log("All submitted jobs have failed! Exiting.")
-                    return None
-                logger.info("Waiting for a job submission (sync + sbatch) to finish...")
-                await asyncio.sleep(1)
-
-            first_running_job, first_state = wait_result
-
-            console.log(
-                f"Job {first_running_job.job_id} on cluster {first_running_job.cluster} is {first_state}. "
-                f"Cancelling the other jobs...\n",
-            )
-            cancelling = True
-            await wait_for_jobs_to_cancel(
-                job_to_state,
-                first_running_job,
-                cluster_to_remote,
-                max_wait_time_seconds,
-            )
-            live.update(make_table(), refresh=True)  # probably not entirely necessary.
-
-        console.print(
-            f"Successfully cancelled all other jobs except for job {first_running_job.job_id} "
-            f"on cluster {first_running_job.cluster}, which is {first_state}."
+    if result.returncode != 0:
+        raise JobSubmissionFailed(
+            f"Failed to submit job on cluster {submission.cluster}: "
+            f"{result.stderr or result.stdout}"
         )
-        if first_state.startswith("RUNNING"):
-            console.print(
-                f"Use `ssh {first_running_job.cluster} sacct -j {first_running_job.job_id}` to view its status."
-            )
-            console.print(
-                f"Once completed, run `cluv sync {first_running_job.cluster}` to fetch its results."
-            )
 
-    except (KeyboardInterrupt, asyncio.CancelledError, Exception):
-        console.log("Interrupted by user. Cancelling all jobs...")
-        to_cancel = list(job_to_state.keys())
-        if first_running_job:
-            to_cancel.remove(first_running_job)
-        await asyncio.gather(
-            *[
-                cancel_job(cluster_to_remote[job.cluster], job.job_id, print=True)
-                for job in to_cancel
-            ]
-        )
-        return None
-
-    return first_running_job
+    return Job(
+        cluster=submission.cluster,
+        remote=submission.remote,
+        job_script=submission.job_script,
+        sbatch_args=submission.sbatch_args,
+        program_args=submission.program_args,
+        sbatch_command=submission.sbatch_command,
+        n_chunks=submission.n_chunks,
+        git_commit=submission.git_commit,
+        job_id=int(result.stdout.strip()),
+        submitted_at=datetime.datetime.now(),
+    )
 
 
 def build_submit_command(
     cluster: str,
-    job_script: str | Path | PurePosixPath,
+    job_script: str | Path | PurePosixPath | None,
     sbatch_args: list[str],
     program_args: list[str],
 ) -> str:
     """Build the local `cluv submit` command line used to launch the job."""
-    command_parts = ["cluv", "submit"]
-    command_parts.extend([cluster, str(job_script), *sbatch_args])
+    command_parts = ["cluv", "submit", cluster]
+    if job_script is not None:
+        command_parts.append(str(job_script))
+    command_parts.extend(sbatch_args)
     if program_args:
         command_parts.extend(["--", *program_args])
     return shlex.join(command_parts)
@@ -610,176 +645,3 @@ def ensure_clean_git_state(autocommit: bool = False, submit_command: str | None 
 
     # Capture current commit hash.
     return subprocess.check_output(["git", "rev-parse", "HEAD"], text=True).strip()
-
-
-def get_job_script_path_from_config(cluster: str) -> Path | PurePosixPath | None:
-    job_script_path = get_cluv_config().get_cluster_config(cluster).job_script_path
-    if cluster == current_cluster() and job_script_path is not None:
-        # Resolve the path to the job script on the local machine.
-        job_script_path = Path(os.path.expandvars(job_script_path))
-        return job_script_path
-    return job_script_path
-
-
-def _check_job_script_not_none(
-    job_script: Path | PurePosixPath | None | None, cluster: str
-) -> Path | PurePosixPath:
-    if job_script is None:
-        raise ValueError(
-            f"No job script was provided and no [tool.cluv] job_script_path is configured for {cluster}."
-        )
-    return job_script
-
-
-def _check_job_script_exists_locally(
-    job_script: Path | PurePosixPath | None, cluster: str
-) -> Path:
-    job_script = _check_job_script_not_none(job_script, cluster)
-    job_script = Path(os.path.expandvars(job_script))
-    if not job_script.exists():
-        raise ValueError(
-            f"The configured job_script value ({job_script}) does not exist on this machine.\n"
-            f"The job script, even though it can be customized per cluster, needs to exist on "
-            f"the local machine, because we need to read its header to infer the values of "
-            f"sbatch arguments."
-        )
-    return job_script
-
-
-def get_sbatch_command(
-    cluster: str,
-    job_script: Path,
-    sbatch_args: list[str],
-    program_args: list[str],
-    git_commit: str,
-    chunking: int | None,
-) -> tuple[str, ResolvedSbatchArgs]:
-    """
-    Generate the command to submit the job via sbatch on the remote cluster, with the appropriate
-    sbatch_arguments, environment variables and paths set.
-
-    NOTE: `sbatch_args` needs to already contain the sbatch arguments from the cluster config + command-line.
-    """
-    # Resolve remote job script path.
-    local_job_script = job_script
-    local_project_dir = find_pyproject().parent
-    if not local_job_script.is_absolute():
-        local_job_script = local_project_dir / local_job_script
-    job_script_relative_path = local_job_script.relative_to(local_project_dir)
-
-    # The project either has a project_dir set, or it is assumed to be under $HOME.
-    remote_project_dir = get_cluv_config().get_cluster_config(cluster).project_dir or (
-        PurePosixPath("$HOME") / local_project_dir.relative_to(Path.home())
-    )
-    remote_job_script = PurePosixPath(remote_project_dir) / job_script_relative_path
-
-    # Build env var dict: global SBATCH_* defaults merged with per-cluster overrides.
-    config = get_cluv_config()
-    cluster_config = config.get_cluster_config(cluster)
-    env_vars: dict[str, str] = {**config.env}
-    env_vars.update(cluster_config.env)
-
-    # Prefix the job name with "cluv-" so it is easy to identify cluv-submitted jobs in sacct.
-    base_name = env_vars.get("SBATCH_JOB_NAME") or Path(job_script).stem
-    env_vars["SBATCH_JOB_NAME"] = f"cluv-{base_name}"
-    env_vars["GIT_COMMIT"] = git_commit
-
-    in_job_packing = False
-    assert not in_job_packing, "todo"
-    # might contain unresolved env vars.
-    cluster_results_path = PurePosixPath(cluster_config.results_path)
-    n_chunks = None
-    # TODO: Use the `get_run_id` function with the placeholder job id %j and task index %t:
-    if chunking:
-        assert not in_job_packing, "can't do both right now."
-        env_vars["SBATCH_OUTPUT"] = f"{cluster_results_path}/{cluster}_%A/slurm-%A_%a.out"
-        n_chunks = get_n_chunks(sbatch_args, env_vars, job_script, chunk_size=chunking)
-        sbatch_args = chunking_update_sbatch_args(n_chunks, sbatch_args, chunk_size=chunking)
-    elif not any("--output" in flag for flag in sbatch_args):
-        if in_job_packing:
-            env_vars["SBATCH_OUTPUT"] = f"{cluster_results_path}/{cluster}_%j_%t/slurm-%j_%t.out"
-        else:
-            env_vars["SBATCH_OUTPUT"] = f"{cluster_results_path}/{cluster}_%j/slurm-%j.out"
-    output_from_cluv = env_vars.get("SBATCH_OUTPUT")
-    if (
-        output_from_file := next(
-            (
-                line
-                for line in job_script.read_text().splitlines()
-                if line.strip().startswith("#SBATCH") and "--output" in line
-            ),
-            None,
-        )
-    ) and output_from_file != output_from_cluv:
-        logger.warning(
-            UserWarning(
-                f"[yellow]⚠️ The job script {job_script} contains an SBATCH --output directive "
-                f"which will be overwritten by cluv, to facilitate the syncing of results.\n"
-                f"Consider using cluv in your Python script to decide where to store results. "
-                f"Take a look a the pytorch example of the Cluv repo for more info.[/yellow]"
-            )
-        )
-
-    env_vars_prefix = " ".join(f"{k}={shlex.quote(str(v))}" for k, v in env_vars.items())
-    sbatch_args_str = shlex.join(sbatch_args)
-    program_args_str = shlex.join(program_args)
-
-    sbatch_command = (
-        f"bash --login -c '{env_vars_prefix} sbatch --parsable --chdir={remote_project_dir} "
-        f"{sbatch_args_str} {remote_job_script} {program_args_str}'"
-    )
-
-    return sbatch_command, ResolvedSbatchArgs(sbatch_args=sbatch_args, n_chunks=n_chunks)
-
-
-async def sbatch(
-    remote: Remote | None,
-    job_script: Path,
-    sbatch_args: list[str],
-    program_args: list[str],
-    git_commit: str,
-    chunking: int | None,
-) -> tuple[subprocess.CompletedProcess[str], Job | None]:
-    """Submit the job via sbatch on the remote cluster, and return the command output and the job info."""
-    cluster = remote.hostname if remote else current_cluster()
-    # Should be set, since `remote` is None if current_cluster() is the same as the cluster argument
-    # to `submit`.
-    assert cluster
-    sbatch_command, resolved_args = get_sbatch_command(
-        cluster=cluster,
-        job_script=job_script,
-        sbatch_args=sbatch_args,
-        program_args=program_args,
-        git_commit=git_commit,
-        chunking=chunking,
-    )
-
-    display = display_commands.get()
-    hide = not display
-    warn = not raise_on_command_error.get()
-
-    if remote:
-        result = await remote.run(sbatch_command, display=display, warn=warn, hide=hide)
-    else:
-        # Run the sbatch command locally.
-        result = await run(
-            tuple(shlex.split(sbatch_command)), _display=display, warn=warn, hide=hide
-        )
-
-    submit_time = datetime.datetime.now()
-
-    if result.returncode != 0:
-        return result, None
-
-    job = Job(
-        job_id=int(result.stdout.strip()),
-        cluster=cluster,
-        job_script=str(job_script),
-        git_commit=git_commit,
-        sbatch_args=resolved_args.sbatch_args,
-        program_args=program_args,
-        submitted_at=submit_time.isoformat(),
-        n_chunks=resolved_args.n_chunks,
-    )
-
-    return result, job
