+++ tests/test_sync_shared_filesystem.py
@@ -7,10 +7,10 @@ jobs through `trillium`. So:
 
 - `get_active_remotes()` must still return a `Remote` for *each* of them.
 - `sync()` must still return a `Remote` for each of them (so job submission can target either).
-- But `_build_sync_tasks()` -- the piece of `sync()` that decides what actually gets synced --
-  must only build one task per shared-filesystem group, since it's the exact same on-disk
-  checkout -- running it twice is redundant, and running it concurrently is unsafe (this is what
-  made `checkout -B <branch> FETCH_HEAD` fail intermittently in the `cluster=first` hydra
+- But `_remotes_to_actually_sync()` -- the piece of `sync()` that decides what actually gets
+  synced -- must only keep one remote per shared-filesystem group, since it's the exact same
+  on-disk checkout -- running it twice is redundant, and running it concurrently is unsafe (this
+  is what made `checkout -B <branch> FETCH_HEAD` fail intermittently in the `cluster=first` hydra
   launcher integration test).
 """
 
@@ -40,23 +40,6 @@ def test_remotes_to_actually_sync_keeps_only_the_first_per_group(hostnames, expe
     remotes = [Remote(hostname=h) for h in hostnames]
     result = sync_module._remotes_to_actually_sync(remotes)
     assert [r.hostname for r in result] == expected
-
-
-def test_build_sync_tasks_skips_remotes_sharing_a_filesystem():
-    """One task per remote, except `trillium-gpu` shares a filesystem with `trillium` -- and thus
-    gets no task of its own -- while unrelated clusters are untouched.
-    """
-    remotes = [
-        Remote(hostname="trillium"),
-        Remote(hostname="trillium-gpu"),
-        Remote(hostname="narval"),
-    ]
-
-    remotes_to_sync, tasks, task_descriptions = sync_module._build_sync_tasks(remotes, here=None)
-
-    assert [r.hostname for r in remotes_to_sync] == ["trillium", "narval"]
-    assert [task.keywords["remote"].hostname for task in tasks] == ["trillium", "narval"]
-    assert task_descriptions == ["local -> trillium", "local -> narval"]
 
 
 async def test_get_active_remotes_returns_every_active_cluster(monkeypatch: pytest.MonkeyPatch):
@@ -117,8 +100,8 @@ async def test_sync_only_syncs_once_but_returns_every_remote(monkeypatch: pytest
     return both remotes.
 
     The dedup logic itself (which remote "wins" a shared-filesystem group) is already covered by
-    `test_build_sync_tasks_skips_remotes_sharing_a_filesystem`, so here we only need to check
-    that `sync()` wires that decision through correctly, by mocking `sync_task_function` as a
+    `test_remotes_to_actually_sync_keeps_only_the_first_per_group`, so here we only need to check
+    that `sync()` wires that decision through correctly, by mocking `sync_per_cluster_part` as a
     single unit rather than each of its internal steps.
     """
     # sync() skips the "is HEAD up to date" check (and thus _head_is_up_to_date) entirely when
@@ -172,8 +155,10 @@ async def test_sync_only_syncs_once_but_returns_every_remote(monkeypatch: pytest
     )
     monkeypatch.setattr(
         sync_module,
-        sync_module.sync_task_function.__name__,
-        mock_sync_task_function := mock.AsyncMock(sync_module.sync_task_function, return_value=[]),
+        sync_module.sync_per_cluster_part.__name__,
+        mock_sync_per_cluster_part := mock.AsyncMock(
+            sync_module.sync_per_cluster_part, return_value=[]
+        ),
     )
 
     remotes = await sync_module.sync(["trillium", "trillium-gpu"], sync_datasets=False)
@@ -181,14 +166,16 @@ async def test_sync_only_syncs_once_but_returns_every_remote(monkeypatch: pytest
     # Both clusters are returned (they're genuinely different Slurm targets)...
     assert {r.hostname for r in remotes} == {"trillium", "trillium-gpu"}
     # ...but the actual sync work only ran once, for the cluster that "wins" the shared group.
-    mock_sync_task_function.assert_awaited_once()
-    assert mock_sync_task_function.await_args.kwargs["remote"].hostname == "trillium"
+    mock_sync_per_cluster_part.assert_awaited_once()
+    assert mock_sync_per_cluster_part.await_args.args[0].hostname == "trillium"
 
     # And every collaborator mocked to get there was actually exercised -- otherwise mocking it
     # would be pointless.
     mock_get_active_remotes.assert_awaited_once()
-    mock_login.assert_awaited_once_with(["trillium", "trillium-gpu"], disabled={})
+    mock_login.assert_awaited_once_with(["trillium", "trillium-gpu"])
     mock_head_is_up_to_date.assert_awaited_once()
     mock_get_cluv_config.assert_called()
-    mock_current_cluster.assert_called_once()
+    # `sync()` resolves `current_cluster()` once itself and once more inside
+    # `sync_common_part()` (each is independently self-contained); not exactly once, but exercised.
+    mock_current_cluster.assert_called()
     mock_get_disabled_clusters.assert_called_once()
